Indicators: Refactoring workers folder [REFACTOR]
