+++ aquitania/indicator/ema.py
