--- aquitania/indicator/workers/__init__.py
@@ -1,20 +1,0 @@
-########################################################################################################################
-# |||||||||||||||||||||||||||||||||||||||||||||||||| AQUITANIA ||||||||||||||||||||||||||||||||||||||||||||||||||||||| #
-# |||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||| #
-# |||| To be a thinker means to go by the factual evidence of a case, not by the judgment of others |||||||||||||||||| #
-# |||| As there is no group stomach to digest collectively, there is no group mind to think collectively. |||||||||||| #
-# |||| Each man must accept responsibility for his own life, each must be sovereign by his own judgment. ||||||||||||| #
-# |||| If a man believes a claim to be true, then he must hold to this belief even though society opposes him. ||||||| #
-# |||| Not only know what you want, but be willing to break all established conventions to accomplish it. |||||||||||| #
-# |||| The merit of a design is the only credential that you require. |||||||||||||||||||||||||||||||||||||||||||||||| #
-# |||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||| #
-########################################################################################################################
-
-"""
-.. moduleauthor:: H Roark
-"""
-
-from .bollinger_bands import BollingerBands
-from .volume import Volume
-from .ema import EMA
-from .rsi import RSI
